--- VaccineCard/VaccineCard/ViewControllers/ScannerViewController.swift
@@ -1,286 +1,0 @@
-//
-//  ScanViewController.swift
-//  VaccineCard
-//
-//  Created by Amir Shayegh on 2021-08-25.
-//
-
-import Foundation
-import AVFoundation
-import UIKit
-
-class ScannerViewController: UIViewController, AVCaptureMetadataOutputObjectsDelegate {
-    // MARK: Constants
-    let flashOnIcon = UIImage(named: "flashOn")
-    let flashOffIcon = UIImage(named: "flashOff")
-    
-    // MARK: Variables
-    private var captureSession: AVCaptureSession?
-    private var previewLayer: AVCaptureVideoPreviewLayer!
-    fileprivate var codeHighlightTags: [Int] = []
-    fileprivate var invalidScannedCodes: [String] = []
-    
-    // Hide Statusbar
-    override var prefersStatusBarHidden: Bool {
-        return true
-    }
-    
-    // Lock in portrait mode
-    override var supportedInterfaceOrientations: UIInterfaceOrientationMask {
-        return .portrait
-    }
-
-    // MARK: LifeCycle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = UIColor.black
-        setupCaptureSession()
-        addFlashlightButton()
-    }
-    
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-
-        if (captureSession?.isRunning == false) {
-            captureSession?.startRunning()
-            setFlash(on: false)
-        }
-    }
-
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-
-        if (captureSession?.isRunning == true) {
-            pauseCamera()
-        }
-    }
-    
-    // MARK: Setup
-    private func setupCaptureSession() {
-        let captureSession = AVCaptureSession()
-        self.captureSession = captureSession
-
-        guard let videoCaptureDevice = AVCaptureDevice.default(for: .video) else { return }
-        
-        // Setup Video input
-        let videoInput: AVCaptureDeviceInput
-
-        do {
-            videoInput = try AVCaptureDeviceInput(device: videoCaptureDevice)
-        } catch {
-            return
-        }
-
-        if (captureSession.canAddInput(videoInput)) {
-            captureSession.addInput(videoInput)
-        } else {
-            alert(title: Constants.Strings.Errors.VideoNotSupported.title,
-                  message: Constants.Strings.Errors.VideoNotSupported.message)
-            return
-        }
-
-        // Setup medatada delegate to capture QR codes
-        let metadataOutput = AVCaptureMetadataOutput()
-        if (captureSession.canAddOutput(metadataOutput)) {
-            captureSession.addOutput(metadataOutput)
-            metadataOutput.setMetadataObjectsDelegate(self, queue: DispatchQueue.main)
-            metadataOutput.metadataObjectTypes = [.qr]
-        } else {
-            alert(title: Constants.Strings.Errors.QRScanningNotSupported.title,
-                  message: Constants.Strings.Errors.QRScanningNotSupported.message)
-            return
-        }
-
-        // Setup Preview
-        previewLayer = AVCaptureVideoPreviewLayer(session: captureSession)
-        previewLayer.frame = view.layer.bounds
-        previewLayer.videoGravity = .resizeAspectFill
-        view.layer.addSublayer(previewLayer)
-
-        // Begin Capture Session
-        captureSession.startRunning()
-    }
-
-    
-    /// Medatada Delegate function - called when a QR is found
-    func metadataOutput(_ output: AVCaptureMetadataOutput, didOutput metadataObjects: [AVMetadataObject], from connection: AVCaptureConnection) {
-        // Remove boxes for previous qr codes
-        clearQRCodeLocations()
-        // if there are multiple codes in camera view
-        if metadataObjects.count > 1 {
-            showMultipleQRCodesWarning(metadataObjects: metadataObjects)
-            return
-        }
-        
-        // get data from single code in view
-        guard let metadataObject = metadataObjects.first,
-              let readableObject = metadataObject as? AVMetadataMachineReadableCodeObject,
-              let stringValue = readableObject.stringValue
-              else {
-            return
-        }
-        
-        // if code has been invalidated already in this session, avoid blocking the camera
-        if !invalidScannedCodes.contains(stringValue) {
-            // Pause camera
-            pauseCamera()
-            // Show code location
-            showQRCodeLocation(for: metadataObject, isInValid: false, tag: Constants.UI.QRCodeHighlighter.tag)
-            // Feedback
-            AudioServicesPlaySystemSound(SystemSoundID(kSystemSoundID_Vibrate))
-            // Validate QR code
-            validate(code: stringValue)
-        } else {
-            // Show message
-            self.showBanner(message: Constants.Strings.Errors.InvalidCode.message)
-            // Show code location
-            showQRCodeLocation(for: metadataObject, isInValid: false, tag: Constants.UI.QRCodeHighlighter.tag)
-        }
-    }
-    
-    fileprivate func validate(code: String) {
-        hideBanner()
-        view.startLoadingIndicator()
-        // Validate
-        CodeValidationService.shared.validate(code: code) { [weak self] result in
-            guard let `self` = self else {return}
-            DispatchQueue.main.async {
-                self.view.endLoadingIndicator()
-                guard let data = result.result else {
-                    // show an error & start camera
-                    switch result.status {
-                    case .ValidCode:
-                        break
-                    case .InvalidCode:
-                        self.showBanner(message: Constants.Strings.Errors.InvalidCode.message)
-                    case .ForgedCode:
-                        self.showBanner(message: Constants.Strings.Errors.ForgedCode.message)
-                    case .MissingData:
-                        self.showBanner(message: Constants.Strings.Errors.IncompleteDataInCode.message)
-                    }
-                    self.startCamera()
-                    self.invalidScannedCodes.append(code)
-                    return
-                }
-                self.found(card: data)
-            }
-        }
-    }
-    
-    public func startCamera() {
-        clearQRCodeLocations()
-        captureSession?.startRunning()
-    }
-    
-    public func pauseCamera() {
-        setFlash(on: false)
-        captureSession?.stopRunning()
-    }
-    
-    /// Called when a SMART QR code is found - override this function
-    /// - Parameter code: QR code
-    func found(card: ScanResultModel) {}
-    
-    fileprivate func showMultipleQRCodesWarning(metadataObjects: [AVMetadataObject]) {
-        for (index, item) in metadataObjects.enumerated() {
-            showQRCodeLocation(for: item, isInValid: true, tag: 1000 + index)
-        }
-        showBanner(message: "There are multiple QR codes in view")
-    }
-    
-    fileprivate func showQRCodeLocation(for object: AVMetadataObject, isInValid: Bool, tag: Int) {
-        guard let metadataLocation = previewLayer.transformedMetadataObject(for: object) else {
-            return
-        }
-        if let existing = view.viewWithTag(tag) {
-            existing.removeFromSuperview()
-        }
-        let container = UIView(frame: metadataLocation.bounds)
-        container.tag = tag
-        container.layer.borderWidth =  Constants.UI.QRCodeHighlighter.borderWidth
-        container.layer.borderColor = isInValid ? Constants.UI.QRCodeHighlighter.borderColorInvalid : Constants.UI.QRCodeHighlighter.borderColor
-        container.layer.cornerRadius =  Constants.UI.QRCodeHighlighter.cornerRadius
-        container.backgroundColor = .clear
-        
-        codeHighlightTags.append(tag)
-        view.addSubview(container)
-        
-        // If its a known invalid QR code, make show invalid colour
-        guard let readableObject = object as? AVMetadataMachineReadableCodeObject,
-              let stringValue = readableObject.stringValue,
-              invalidScannedCodes.contains(stringValue)
-              else {
-            return
-        }
-        container.layer.borderColor = Constants.UI.QRCodeHighlighter.borderColorInvalid
-    }
-    
-    fileprivate func clearQRCodeLocations() {
-        for tag in codeHighlightTags {
-            if let box = view.viewWithTag(tag) {
-                box.removeFromSuperview()
-            }
-        }
-    }
-    
-    func setFlash(on: Bool) {
-        guard
-            let device = AVCaptureDevice.default(for: AVMediaType.video),
-            device.hasTorch
-        else { return }
-
-        do {
-            try device.lockForConfiguration()
-            device.torchMode = on ? .on : .off
-            device.unlockForConfiguration()
-        } catch {
-            print("Flash could not be used")
-        }
-        
-        // TODO: TAG FROM CONSTANT
-        guard let btn = self.view.viewWithTag(92133) as? UIButton else {
-            return
-        }
-        if on {
-            btn.setImage(flashOnIcon, for: .normal)
-        } else {
-            btn.setImage(flashOffIcon, for: .normal)
-        }
-    }
-    
-    fileprivate func addFlashlightButton() {
-        // TODO: Refactor constants
-        let btnSize: CGFloat = 42
-        let button = UIButton(frame: CGRect(x: 0, y: 0, width: btnSize, height: btnSize))
-        button.tag = 92133
-        view.addSubview(button)
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.heightAnchor.constraint(equalToConstant: btnSize).isActive = true
-        button.widthAnchor.constraint(equalToConstant: btnSize).isActive = true
-        button.topAnchor.constraint(equalTo: view.topAnchor, constant: 32).isActive = true
-        button.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16).isActive = true
-        button.backgroundColor = .lightGray
-        button.setImage(flashOffIcon, for: .normal)
-        
-        button.addTarget(self, action: #selector(flashTapped), for: .touchUpInside)
-        button.layer.cornerRadius = btnSize/2
-        
-        button.imageView?.contentMode = .scaleAspectFit
-        button.imageEdgeInsets = UIEdgeInsets(top: 8, left: 8, bottom: 8, right: 8)
-    }
-    
-    @objc func flashTapped(sender: UIButton?) {
-        // TODO: Refactor constants (Tag)
-        guard let btn = self.view.viewWithTag(92133) as? UIButton else {
-            return
-        }
-        let isOn = btn.imageView?.image == flashOnIcon
-        if isOn {
-            setFlash(on: false)
-        } else {
-            setFlash(on: true)
-        }
-    }
-    
-    
-}
